add default get_batch func
