+++ tools/train.sh
@@ -8,5 +8,5 @@ ADDR=${ADDR:-127.0.0.1}
 
 python3 -m oneflow.distributed.launch \
 --nproc_per_node $GPUS --nnodes $NODE --node_rank $NODE_RANK --master_addr $ADDR \
-tools/pretrain_bert.py \
+tools/train_net.py \
 --config-file $CONFIG ${@:3}
